requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.13.3
aiohttp==3.9.3
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.6.3
transformers==4.38.2
torch
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
httpx==0.27.0
wikipedia==1.4.0
python-multipart==0.0.9
typing-extensions==4.9.0